requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
WTForms==3.1.1
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==3.0.1
email-validator==2.1.0
gunicorn==21.2.0
